fix: remove 함수 고침

+++ Assignment1/pomoku.c
@@ -318,14 +318,14 @@ int insert_column(const color_t color, const size_t col)
 
 int remove_row(const color_t color, const size_t row)
 {
-    size_t i;
+    int i;
     size_t col;
     
     if (is_removable_row(row) == FALSE || is_valid_score(color, INSERT_SCORES) == FALSE) {
         return FALSE;
     }
 
-    for (i = row; i < g_row; i++) {
+    for (i = row; i < (int)g_row; i++) {
         for (col = 0; col < g_col; col++) {
             g_board[i - 1][col] = g_board[i][col];
         }
@@ -343,14 +343,14 @@ int remove_row(const color_t color, const size_t row)
 
 int remove_column(const color_t color, const size_t col)
 {
-    size_t i;
+    int i;
     size_t row;
 
     if (is_removable_col(col) == FALSE || is_valid_score(color, INSERT_SCORES) == FALSE) {
         return FALSE;
     }
 
-    for (i = col; i < g_col; i++) {
+    for (i = col; i < (int)g_col; i++) {
         for (row = 0; row < g_row; row++) {
             g_board[row][i - 1] = g_board[row][i];
         }
